requirements.txt
tensorflow
flask
Pillow
numpy
